+++ ADMF/functions/Invoke-AdmfDC.ps1
@@ -43,6 +43,9 @@
 		
 		[PSCredential]
 		$Credential,
+
+        [string[]]
+        $TargetServer = @(),
 		
 		[ADMF.UpdateDCOptions[]]
 		$Options = 'Default',
@@ -59,6 +62,9 @@
 	{
 		Reset-DomainControllerCache
 		$parameters = $PSBoundParameters | ConvertTo-PSFHashtable -Include Server, Credential
+        if (-not $Server -and $TargetServer) {
+            $parameters.Server = $TargetServer | Select-Object -First 1
+        }
 		$originalArgument = Invoke-PreCredentialProvider @parameters -ProviderName $CredentialProvider -Parameter $parameters -Cmdlet $PSCmdlet
 		try { $dcServer = Resolve-DomainController @parameters -Confirm:$false }
 		catch
@@ -82,7 +88,7 @@
 				if (Get-DCShare)
 				{
 					Write-PSFMessage -Level Host -String 'Invoke-AdmfDC.Executing.Invoke' -StringValues 'Shares', $parameters.Server
-					Invoke-DCShare @parameters
+					Invoke-DCShare @parameters -TargetServer $TargetServer
 				}
 				else { Write-PSFMessage -Level Host -String 'Invoke-AdmfDC.Skipping.Test.NoConfiguration' -StringValues 'Shares' }
 			}
@@ -91,7 +97,7 @@
 				if (Get-DCAccessRule)
 				{
 					Write-PSFMessage -Level Host -String 'Invoke-AdmfDC.Executing.Invoke' -StringValues 'FSAccessRules', $parameters.Server
-					Invoke-DCAccessRule @parameters
+					Invoke-DCAccessRule @parameters -TargetServer $TargetServer
 				}
 				else { Write-PSFMessage -Level Host -String 'Invoke-AdmfDC.Skipping.Test.NoConfiguration' -StringValues 'FSAccessRules' }
 			}